New model for Views: simply convert between normalized and rectified coordinates
Trying to make a calibrator that goes between distorted and rectified *pixel*
coordinates in a single object turned out to be super-cludgy. There was an
annoying amount of bookkeeping (Calibration expects this so View X should do
that). By splitting the previous notion of a view into two parts, things
become much simpler. A single view can only convert between pixel coordinates
and normalized coordinates (e.g. X/Z, Y/Z for a 3D point X,Y,Z). For
CameraCalibrations, this transformation includes applying or removing
distortion, as appropriate. The advantage of this is that transformations
can be easily applied to explain both the acquired image and the desired
image, even though they may look very different. In addition, the underlying
code in the view transforms is much simpler.

It now takes two views (e.g. calibration and a max rectified view) to convert
an image from the camera with a rasterizer. The rasterizer can also handle
rotations of the camera as long as both extrinsics matrices (before and after)
are known.