Initialize the mosaic extrinsics with a joint homography and homography decomposition
Fit a homography to all of the tag observations in a frame instead of using
per-tag homographies. Decompose this homography with a method like
CameraUtil.homographyToPose, but with fewer assumptions and a different
coordinate convention -- CameraMath uses the same one as Caltech (Z points
into the scene)

Also, CameraUtil.homographyToPose has a typo in the math in the comments that
never seems to go away. Fixed (again)